refactor: auth 패키지에서만 의존성이 있는 클래스를 하위로 이동
